small file structure refactor
