+++ api/schemas/shared_structures/__init__.py
@@ -1,0 +1,2 @@
+from schemas.shared_structures.auth_result import AuthResult
+from schemas.shared_structures.category_totals import CategoryTotals
