+++ api/schemas/shared_structures/auth_result.py
