+++ api/schemas/shared_structures/category_totals.py
